Add support for Debian version 13 codename Trixie (#451)

+++ debian/resources/switch/source-release.sh
@@ -12,21 +12,27 @@ apt update && apt upgrade -y
 
 # install dependencies
 apt install -y autoconf automake devscripts g++ git-core libncurses5-dev libtool libtool-bin make libjpeg-dev
-apt install -y pkg-config flac  libgdbm-dev libdb-dev gettext sudo equivs mlocate git dpkg-dev libpq-dev
+apt install -y pkg-config flac  libgdbm-dev libdb-dev gettext sudo equivs git dpkg-dev libpq-dev
 apt install -y liblua5.2-dev libtiff5-dev libperl-dev libcurl4-openssl-dev libsqlite3-dev libpcre3-dev
 apt install -y devscripts libspeexdsp-dev libspeex-dev libldns-dev libedit-dev libopus-dev libmemcached-dev
 apt install -y libshout3-dev libmpg123-dev libmp3lame-dev yasm nasm libsndfile1-dev libuv1-dev libvpx-dev
-apt install -y libavformat-dev libswscale-dev libvlc-dev python3-distutils sox libsox-fmt-all
+apt install -y libavformat-dev libswscale-dev libvlc-dev sox libsox-fmt-all
 
 #install dependencies that depend on the operating system version
+if [ ."$os_codename" = ."noble" ]; then
+	apt install -y python3-distutils mlocate libvpx9 swig3.0
+fi
 if [ ."$os_codename" = ."stretch" ]; then
-	apt install -y libvpx4 swig3.0
+	apt install -y python3-distutils mlocate libvpx4 swig3.0
 fi
 if [ ."$os_codename" = ."buster" ]; then
-	apt install -y libvpx5 swig3.0
+	apt install -y python3-distutils mlocate libvpx5 swig3.0
 fi
 if [ ."$os_codename" = ."bullseye" ]; then
-	apt install -y libvpx6 swig4.0
+	apt install -y python3-distutils mlocate libvpx6 swig4.0
+fi
+if [ ."$os_codename" = ."trixie" ]; then
+	apt install -y python3-distutils-extra plocate
 fi
 
 # additional dependencies
@@ -36,7 +42,7 @@ apt install -y sqlite3 unzip
 CWD=$(pwd)
 
 #install the following dependencies if the switch version is greater than 1.10.0
-if [ $(echo "$switch_version" | tr -d '.') -gt 1100 ]; then
+if [ ."$switch_version" = ."master" ] || [ $(echo "$switch_version" | tr -d '.') -gt 1100 ] || [ ."$os_codename" = ."trixie" ]; then
 
 	# libks build-requirements
 	apt install -y cmake uuid-dev
@@ -54,10 +60,14 @@ if [ $(echo "$switch_version" | tr -d '.') -gt 1100 ]; then
 
 	# sofia-sip
 	cd /usr/src
-	#git clone https://github.com/freeswitch/sofia-sip.git sofia-sip
-	wget https://github.com/freeswitch/sofia-sip/archive/refs/tags/v$sofia_version.zip
-	unzip v$sofia_version.zip
-	cd sofia-sip-$sofia_version
+	if [ ."$sofia_version" = ."master" ] || [ ."$os_codename" = ."trixie" ]; then
+		git clone https://github.com/freeswitch/sofia-sip.git sofia-sip
+		cd sofia-sip
+	else
+		wget https://github.com/freeswitch/sofia-sip/archive/refs/tags/v$sofia_version.zip
+		unzip v$sofia_version.zip
+		cd sofia-sip-$sofia_version
+	fi
 	sh autogen.sh
 	./configure --enable-debug
 	make -j $(getconf _NPROCESSORS_ONLN)
@@ -67,8 +77,10 @@ if [ $(echo "$switch_version" | tr -d '.') -gt 1100 ]; then
 	cd /usr/src
 	git clone https://github.com/freeswitch/spandsp.git spandsp
 	cd spandsp
- 	git reset --hard 0d2e6ac65e0e8f53d652665a743015a88bf048d4
- 	#/usr/bin/sed -i 's/AC_PREREQ(\[2\.71\])/AC_PREREQ([2.69])/g' /usr/src/spandsp/configure.ac
+	if [ ."$sofia_version" != ."master" ] && [ ."$os_codename" != ."trixie" ] && [ ."$switch_branch" != ."master" ]; then
+		git reset --hard 0d2e6ac65e0e8f53d652665a743015a88bf048d4
+	fi
+	#/usr/bin/sed -i 's/AC_PREREQ(\[2\.71\])/AC_PREREQ([2.69])/g' /usr/src/spandsp/configure.ac
 	sh autogen.sh
 	./configure --enable-debug
 	make -j $(getconf _NPROCESSORS_ONLN)
@@ -79,35 +91,48 @@ fi
 cd /usr/src
 
 #check for master
-if [ $switch_branch = "master" ]; then
+if [ ."$os_codename" != ."trixie" ] || [ ."$switch_branch" = ."master" ]; then
 	#master branch
 	echo "Using version master"
 	rm -r /usr/src/freeswitch
 	git clone https://github.com/signalwire/freeswitch.git
 	cd /usr/src/freeswitch
+
+	git remote add fusionpbx https://github.com/fusionpbx/freeswitch.git
+	git fetch fusionpbx
+	git checkout -b $switch_version
+
+	git rebase fusionpbx/master
 	./bootstrap.sh -j
 fi
 
 #check for stable release
-if [ $switch_branch = "stable" ]; then
+if [ ."$switch_branch" != ."master" ] && [ ."$os_codename" != ."trixie" ] && [ ."$switch_branch" = ."stable" ]; then
 	echo "Using version $switch_version"
 	#1.8 and older
 	if [ $(echo "$switch_version" | tr -d '.') -lt 1100 ]; then
 		wget http://files.freeswitch.org/freeswitch-releases/freeswitch-$switch_version.zip
 		unzip freeswitch-$switch_version.zip
 		cd /usr/src/freeswitch-$switch_version
+
+		# Reset repo just-in-case we are rebuilding
+		git reset --hard HEAD && git clean -fdx
 	fi
 
 	#1.10.0 and newer
 	if [ $(echo "$switch_version" | tr -d '.') -gt 1100 ]; then
-		git clone -b $switch_version --single-branch https://github.com/fusionpbx/freeswitch freeswitch-$switch_version
-  		git checkout $switch_version
+
+		git clone https://github.com/fusionpbx/freeswitch freeswitch-$switch_version
+
+		cd /usr/src/freeswitch-$switch_version
+
+		# Reset repo just-in-case we are rebuilding
+		git reset --hard origin/master && git clean -fdx
 
 		#wget http://files.freeswitch.org/freeswitch-releases/freeswitch-$switch_version.-release.zip
 		#unzip freeswitch-$switch_version.-release.zip
 		#mv freeswitch-$switch_version.-release freeswitch-$switch_version
-		cd /usr/src/freeswitch-$switch_version
-		
+
 		# bootstrap is needed if using git
 		./bootstrap.sh -j
 
@@ -124,22 +149,22 @@ fi
 
 # enable required modules
 #sed -i /usr/src/freeswitch/modules.conf -e s:'#applications/mod_avmd:applications/mod_avmd:'
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'#applications/mod_av:formats/mod_av:'
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'#applications/mod_callcenter:applications/mod_callcenter:'
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'#applications/mod_cidlookup:applications/mod_cidlookup:'
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'#applications/mod_memcache:applications/mod_memcache:'
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'#applications/mod_nibblebill:applications/mod_nibblebill:'
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'#applications/mod_curl:applications/mod_curl:'
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'#applications/mod_translate:applications/mod_translate:'
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'#formats/mod_shout:formats/mod_shout:'
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'#formats/mod_pgsql:formats/mod_pgsql:'
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'#say/mod_say_es:say/mod_say_es:'
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'#say/mod_say_fr:say/mod_say_fr:'
+sed -i modules.conf -e s:'#applications/mod_av:formats/mod_av:'
+sed -i modules.conf -e s:'#applications/mod_callcenter:applications/mod_callcenter:'
+sed -i modules.conf -e s:'#applications/mod_cidlookup:applications/mod_cidlookup:'
+sed -i modules.conf -e s:'#applications/mod_memcache:applications/mod_memcache:'
+sed -i modules.conf -e s:'#applications/mod_nibblebill:applications/mod_nibblebill:'
+sed -i modules.conf -e s:'#applications/mod_curl:applications/mod_curl:'
+sed -i modules.conf -e s:'#applications/mod_translate:applications/mod_translate:'
+sed -i modules.conf -e s:'#formats/mod_shout:formats/mod_shout:'
+sed -i modules.conf -e s:'#formats/mod_pgsql:formats/mod_pgsql:'
+sed -i modules.conf -e s:'#say/mod_say_es:say/mod_say_es:'
+sed -i modules.conf -e s:'#say/mod_say_fr:say/mod_say_fr:'
 
 #disable module or install dependency libks to compile signalwire
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'applications/mod_signalwire:#applications/mod_signalwire:'
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'endpoints/mod_skinny:#endpoints/mod_skinny:'
-sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'endpoints/mod_verto:#endpoints/mod_verto:'
+sed -i modules.conf -e s:'applications/mod_signalwire:#applications/mod_signalwire:'
+sed -i modules.conf -e s:'endpoints/mod_skinny:#endpoints/mod_skinny:'
+sed -i modules.conf -e s:'endpoints/mod_verto:#endpoints/mod_verto:'
 
 # prepare the build
 #./configure --prefix=/usr/local/freeswitch --enable-core-pgsql-support --disable-fhs
@@ -150,6 +175,9 @@ sed -i /usr/src/freeswitch-$switch_version/modules.conf -e s:'endpoints/mod_vert
 # compile and install
 make -j $(getconf _NPROCESSORS_ONLN)
 make install
+
+# create voicemail directory for installer
+mkdir -p /var/lib/freeswitch/storage/voicemail
 
 #return to the executing directory
 cd $CWD
